(PC-15100) fix missing navigation object crash
