+++ src/features/errors/pages/AsyncErrorBoundary.tsx
@@ -1,26 +1,33 @@
 import { t } from '@lingui/macro'
-import React, { useEffect } from 'react'
+import React, { ReactNode, useEffect } from 'react'
 import { FallbackProps } from 'react-error-boundary'
 import { useQueryErrorResetBoundary } from 'react-query'
 import styled from 'styled-components/native'
 
+import { homeNavConfig } from 'features/navigation/TabBar/helpers'
+import { useGoBack } from 'features/navigation/useGoBack'
 import { AsyncError, MonitoringError, eventMonitoring } from 'libs/monitoring'
 import { ScreenError } from 'libs/monitoring/errors'
 import { Helmet } from 'libs/react-helmet/Helmet'
 import { ButtonPrimaryWhite } from 'ui/components/buttons/ButtonPrimaryWhite'
+import { styledButton } from 'ui/components/buttons/styledButton'
 import { GenericErrorPage } from 'ui/components/GenericErrorPage'
+import { Touchable } from 'ui/components/touchable/Touchable'
 import { BrokenConnection } from 'ui/svg/BrokenConnection'
-import { Typo } from 'ui/theme'
+import { ArrowPrevious } from 'ui/svg/icons/ArrowPrevious'
+import { getSpacing, Typo } from 'ui/theme'
+import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
 
 interface AsyncFallbackProps extends FallbackProps {
   resetErrorBoundary: (...args: Array<unknown>) => void
   error: AsyncError | ScreenError | Error
-  headerGoBack?: boolean
+  header?: ReactNode
 }
 
 export const AsyncErrorBoundaryWithoutNavigation = ({
   resetErrorBoundary,
   error,
+  header,
 }: AsyncFallbackProps) => {
   const { reset } = useQueryErrorResetBoundary()
 
@@ -57,7 +64,7 @@ export const AsyncErrorBoundaryWithoutNavigation = ({
       <GenericErrorPage
         title={t`Oups\u00a0!`}
         icon={BrokenConnection}
-        headerGoBack
+        header={header}
         buttons={[
           <ButtonPrimaryWhite
             key={1}
@@ -73,8 +80,35 @@ export const AsyncErrorBoundaryWithoutNavigation = ({
 }
 
 export const AsyncErrorBoundary = (props: AsyncFallbackProps) => {
-  return <AsyncErrorBoundaryWithoutNavigation {...props} />
+  const { goBack, canGoBack } = useGoBack(...homeNavConfig)
+  const { top } = useCustomSafeInsets()
+
+  return (
+    <AsyncErrorBoundaryWithoutNavigation
+      {...props}
+      header={
+        !!canGoBack() && (
+          <HeaderContainer onPress={goBack} top={top + getSpacing(3.5)} testID="Revenir en arrière">
+            <StyledArrowPrevious />
+          </HeaderContainer>
+        )
+      }
+    />
+  )
 }
+
+const StyledArrowPrevious = styled(ArrowPrevious).attrs(({ theme }) => ({
+  color: theme.colors.white,
+  size: theme.icons.sizes.small,
+  accessibilityLabel: t`Revenir en arrière`,
+}))``
+
+const HeaderContainer = styledButton(Touchable)<{ top: number }>(({ theme, top }) => ({
+  position: 'absolute',
+  top,
+  left: getSpacing(6),
+  zIndex: theme.zIndex.floatingButton,
+}))
 
 const StyledBody = styled(Typo.Body)(({ theme }) => ({
   color: theme.colors.white,
