+++ src/ui/components/GenericErrorPage.tsx
@@ -1,21 +1,14 @@
-import { t } from '@lingui/macro'
 import React, { ReactNode, FunctionComponent } from 'react'
 import styled, { useTheme } from 'styled-components/native'
 
-import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { useGoBack } from 'features/navigation/useGoBack'
 import { Helmet } from 'libs/react-helmet/Helmet'
-import { styledButton } from 'ui/components/buttons/styledButton'
-import { Touchable } from 'ui/components/touchable/Touchable'
 import { Background } from 'ui/svg/Background'
-import { ArrowPrevious } from 'ui/svg/icons/ArrowPrevious'
 import { IconInterface } from 'ui/svg/icons/types'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
 import { getHeadingAttrs } from 'ui/theme/typography'
-import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
 
 type Props = {
-  headerGoBack?: boolean
+  header?: ReactNode
   noIndex?: boolean
   icon?: FunctionComponent<IconInterface>
   title: string
@@ -24,15 +17,13 @@ type Props = {
 
 export const GenericErrorPage: FunctionComponent<Props> = ({
   children,
-  headerGoBack,
+  header,
   noIndex = true,
   icon,
   title,
   buttons,
 }) => {
   const { isTouch } = useTheme()
-  const { top } = useCustomSafeInsets()
-  const { canGoBack, goBack } = useGoBack(...homeNavConfig)
   const Icon =
     !!icon &&
     styled(icon).attrs(({ theme }) => ({
@@ -48,11 +39,7 @@ export const GenericErrorPage: FunctionComponent<Props> = ({
         </Helmet>
       )}
       <Background />
-      {headerGoBack && canGoBack() ? (
-        <HeaderContainer onPress={goBack} top={top + getSpacing(3.5)} testID="Revenir en arrière">
-          <StyledArrowPrevious />
-        </HeaderContainer>
-      ) : null}
+      {header}
       <Content>
         <Spacer.TopScreen />
         <Spacer.Flex />
@@ -97,19 +84,6 @@ const spacingMatrix = {
   afterChildren: 10,
   bottom: 10,
 }
-
-const StyledArrowPrevious = styled(ArrowPrevious).attrs(({ theme }) => ({
-  color: theme.colors.white,
-  size: theme.icons.sizes.small,
-  accessibilityLabel: t`Revenir en arrière`,
-}))``
-
-const HeaderContainer = styledButton(Touchable)<{ top: number }>(({ theme, top }) => ({
-  position: 'absolute',
-  top,
-  left: getSpacing(6),
-  zIndex: theme.zIndex.floatingButton,
-}))
 
 const StyledTitle = styled(Typo.Title2).attrs(() => getHeadingAttrs(1))(({ theme }) => ({
   color: theme.colors.white,
